Edit Form Fix + Indv. Profile takes Contact Obj
There was an error with how the Edit Profile form was made that prevented it from properly linking UI elements to their functions. This has been fixed. Additionally, instead of taking a list widget item as a parameter, the program now finds the name of the matching Contact object and sends that instead.